Controller for tanecne teleso and Sutaz

+++ comma_rest/src/main/java/com/comma/comma_rest/dao/SutazDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface SutazDao {
 
-    Sutaz findById(int id);
+    Sutaz findById(long id);
 
     List<Sutaz> findAll();
